raise a Closure::CompilationError if java exits with an error
Nice to know that an error occurred and even nicer to get the error output from stderr.
This requires switching to popen from popen3 because the latter is broken in ruby 1.8 when it comes to exit codes (http://redmine.ruby-lang.org/issues/show/1287).
Instead, we redirect stderr to a tempfile and read the tempfile contents so we can raise that in the error if one occurs.