Use Closure Compiler online API
This commit makes closure-compiler use the online Closure Compiler API
hosted by Google. Code which was using the Java version of the Closure
Compiler was removed.

All code-compression related tests pass. Others where removed because of
not being applicable after switching to the online API.

Benchmarks on a 32 bit Ubuntu 11.10 show a significant speedup.

    $ git checkout 637abe13b # A version before this patch
    $ time rake test
    ...
    10 tests, 11 assertions, 0 failures, 0 errors, 0 skips
    38.50user 1.48system 0:23.19elapsed 172%CPU
    (0avgtext+0avgdata 611904maxresident)k
    4720inputs+0outputs (8major+203726minor)pagefaults 0swaps

Apply this patch and execute:

    $ time rake test
    ...
    6 tests, 7 assertions, 0 failures, 0 errors, 0 skips
    0.51user 0.05system 0:11.61elapsed 4%CPU
    (0avgtext+0avgdata 61392maxresident)k
    0inputs+0outputs (0major+6388minor)pagefaults 0swaps

Signed-off-by: Jan Stępień <jstepien@users.sourceforge.net>